amazon_api_service/requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confuse==2.0.1
cssselect==1.2.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.0
flower==2.0.1
frozenlist==1.4.1
h11==0.14.0
humanize==4.10.0
idna==3.10
importlib_metadata==8.5.0
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
lxml==5.2.2
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mkl==2024.2.2
mkl-service==2.4.0
multidict==6.1.0
numpy==1.24.4
pandas==1.4.1
parse==1.20.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
pycryptodome==3.20.0
pydantic==2.10.4
pydantic_core==2.27.2
pyee
PyMySQL==1.1.1
pyppeteer==1.0.0
pyquery==2.0.1
python-amazon-ad-api==0.6.0
python-amazon-advertising-api==0.8.2
python-amazon-sp-api==1.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
requests-html==0.10.0
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tbb==2021.13.1
tenacity==9.0.0
tornado==6.4.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
uvicorn==0.34.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.1.3
yarl==1.15.2
zipp==3.21.0
